chore: add to multiselect
